requirements.txt
allure-behave==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
behave==1.2.6
certifi==2025.1.31
colorama==0.4.6
distro==1.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.66.3
parse==1.20.2
parse_type==0.6.4
playwright==1.50.0
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
